requirements.txt
flask
pandas
snowflake-connector-python
azure-functions
numpy==1.26.4
python-dotenv